Primitives

+++ BongusCode_Compiler/src/parser/parser.hpp
@@ -203,7 +203,7 @@ namespace yy {
     {
 #line 45 "parser.y"
 
-	int num;
+	unsigned long long int num;
 	// const wchar_t* str;
 	std::wstring* str;
 	AST::Node* ASTNode;
